requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
feedparser==6.0.10
newspaper3k==0.2.8
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
chromadb==0.4.15
langchain==0.0.335
langchain-community==0.0.5
huggingface-hub==0.17.3
accelerate==0.24.1
bitsandbytes==0.41.1
scipy==1.11.3
scikit-learn==1.3.0
matplotlib==3.7.2
plotly==5.17.0
textstat==0.7.3
nltk==3.8.1
spacy==3.7.2
datasets==2.14.6